Unify filterConfig
