+++ src/handlers/gestureHandlerCommon.ts
@@ -3,9 +3,11 @@
 // e.g. React.createRef<TapGestureHandler> -> React.createRef<typeof TapGestureHandler>.
 // See https://www.typescriptlang.org/docs/handbook/classes.html#constructor-functions for reference.
 import * as React from 'react';
+import { Platform } from 'react-native';
 
 import { State } from '../State';
 import { ValueOf } from '../typeUtils';
+import { handlerIDToTag } from './handlersRegistry';
 
 const commonProps = [
   'id',
@@ -28,6 +30,8 @@ export const baseGestureHandlerProps = [
   'onGestureEvent',
   'onHandlerStateChange',
 ] as const;
+
+export const baseGestureHandlerWithMonitorProps = [...commonProps] as const;
 
 export interface GestureEventPayload {
   handlerTag: number;
@@ -102,3 +106,57 @@ export type BaseGestureHandlerProps<
     event: HandlerStateChangeEvent<ExtraEventPayloadT>
   ) => void;
 };
+
+function isConfigParam(param: unknown, name: string) {
+  // param !== Object(param) returns false if `param` is a function
+  // or an object and returns true if `param` is null
+  return (
+    param !== undefined &&
+    (param !== Object(param) ||
+      !('__isNative' in (param as Record<string, unknown>))) &&
+    name !== 'onHandlerStateChange' &&
+    name !== 'onGestureEvent'
+  );
+}
+
+export function filterConfig(
+  props: Record<string, unknown>,
+  validProps: string[],
+  defaults: Record<string, unknown> = {}
+) {
+  const res = { ...defaults };
+  validProps.forEach((key) => {
+    const value = props[key];
+    if (isConfigParam(value, key)) {
+      let value = props[key];
+      if (key === 'simultaneousHandlers' || key === 'waitFor') {
+        value = transformIntoHandlerTags(props[key]);
+      } else if (key === 'hitSlop') {
+        if (typeof value !== 'object') {
+          value = { top: value, left: value, bottom: value, right: value };
+        }
+      }
+      res[key] = value;
+    }
+  });
+  return res;
+}
+
+function transformIntoHandlerTags(handlerIDs: any) {
+  if (!Array.isArray(handlerIDs)) {
+    handlerIDs = [handlerIDs];
+  }
+
+  if (Platform.OS === 'web') {
+    return handlerIDs
+      .map(({ current }: { current: any }) => current)
+      .filter((handle: any) => handle);
+  }
+  // converts handler string IDs into their numeric tags
+  return handlerIDs
+    .map(
+      (handlerID: any) =>
+        handlerIDToTag[handlerID] || handlerID.current?.handlerTag || -1
+    )
+    .filter((handlerTag: number) => handlerTag > 0);
+}
